101 print number 2

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
 
 void print_number(int n)
 {
-	long num = n;
+	int num = n;
 
 	if (num < 0)
 	{
